refactor: agrupa funções auxiliares relacionadas como métodos em uma classe
